+++ HISTORY.md
@@ -6,6 +6,11 @@ No advantage over using file_get_contents() and parsing contents.
 
 Instead using PHP gererator function to split processing to discrete chunks so that only the current row is in memory.
 
+# 2019-12-30 Check PROP_SUB_ID
+- Add test for PROP_SUB_ID
+- Add test fot int
+- Add test for num
+
 # 2019-12-30 Check date fields
 - add test for date columns created, updated, let-date-available